core/anti_spoofing/train/requirements.txt
torchvision
onnx
onnxruntime
numpy
Pillow
